src/python/requirements.txt
oracledb==3.1.0
SQLAlchemy==2.0.40
streamlit==1.44.1
pandas==2.2.3
numpy==2.2.5
python-dotenv==1.1.0